+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## HEAD
 
-### New features
+### New package features
 
+- Add the implementation of some basic shapes, such as spheres ([#11](https://github.com/Paolo97Gll/Raytracer.jl/pull/11)).
 - Add the implementation of an image tracer, needed to capture light from a scene ([#3](https://github.com/Paolo97Gll/Raytracer.jl/pull/3)).
 - Add geometry implementation, needed to compute a 3D scene ([#2](https://github.com/Paolo97Gll/Raytracer.jl/pull/2)).
+
+### New CLI tool features
+
+- Add `demo` command to show a demo rendering ([#13](https://github.com/Paolo97Gll/Raytracer.jl/pull/13)).
 
 ## v0.1.0